refactor: Moved everything into packages
