+++ src/main/java/duke/command/EventCommand.java
@@ -1,22 +1,35 @@
+package duke.command;
+
 import java.io.IOException;
 import java.time.format.DateTimeParseException;
+import duke.task.Task;
+import duke.task.Todo;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.storage.Storage;
+import duke.storage.TaskList;
+import duke.ui.Ui;
+import duke.exception.DukeException;
 
-public class DeadlineCommand extends Command {
+
+public class EventCommand extends Command {
     private String input;
     
-    public DeadlineCommand(String input) {
+    public EventCommand(String input) {
         this.input = input;
     }
     
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        String[] deadlineParts = input.split(" /by ");
-        String deadlineDesc = deadlineParts[0].substring(8).trim();
-        String by = deadlineParts[1];
+        String[] fromSplit = input.split(" /from ");
+        String eventDesc = fromSplit[0].substring(5).trim();
+        String[] toSplit = fromSplit[1].split(" /to ");
+        String from = toSplit[0];
+        String to = toSplit[1];
         try {
-            Task deadlineTask = new Deadline(deadlineDesc, by);
-            tasks.add(deadlineTask);
-            ui.showTaskAdded(deadlineTask, tasks.size());
+            Task eventTask = new Event(eventDesc, from, to);
+            tasks.add(eventTask);
+            ui.showTaskAdded(eventTask, tasks.size());
             saveTasksToFile(tasks, ui, storage);
         } catch (DateTimeParseException e) {
             ui.showError("OOPS!!! " + e.getMessage());
