+++ src/main/java/duke/exception/DeadlineException.java
@@ -1,3 +1,4 @@
+package duke.exception;
 public class DeadlineException extends DukeException {
     public DeadlineException(String message) {
         super(message);
