+++ src/main/java/duke/exception/DeleteException.java
@@ -1,3 +1,4 @@
+package duke.exception;
 public class DeleteException extends DukeException {
     public DeleteException(String message) {
         super(message);
