+++ src/main/java/duke/exception/EventException.java
@@ -1,3 +1,4 @@
+package duke.exception;
 public class EventException extends DukeException {
     public EventException(String message) {
         super(message);
