+++ src/main/java/duke/exception/InvalidCommandException.java
@@ -1,3 +1,4 @@
+package duke.exception;
 public class InvalidCommandException extends DukeException {
     public InvalidCommandException(String message) {
         super(message);
