+++ src/main/java/duke/exception/InvalidTaskNumberException.java
@@ -1,3 +1,4 @@
+package duke.exception;
 public class InvalidTaskNumberException extends DukeException {
     public InvalidTaskNumberException(String message) {
         super(message);
