+++ src/main/java/duke/exception/TodoException.java
@@ -1,3 +1,4 @@
+package duke.exception;
 public class TodoException extends DukeException {
     public TodoException(String message) {
         super(message);
