+++ src/main/java/duke/storage/Storage.java
@@ -1,3 +1,4 @@
+package duke.storage;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
@@ -5,6 +6,7 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
+import duke.task.Task;
 
 public class Storage {
     private static final String DEFAULT_DATA_DIR = "data";
