+++ src/main/java/duke/storage/TaskList.java
@@ -1,4 +1,6 @@
+package duke.storage;
 import java.util.ArrayList;
+import duke.task.Task;
 
 public class TaskList {
     private ArrayList<Task> tasks;
