+++ src/main/java/duke/task/Event.java
@@ -1,3 +1,4 @@
+package duke.task;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
