+++ src/main/java/duke/task/Todo.java
@@ -1,3 +1,4 @@
+package duke.task;
 import java.io.IOException;
 
 public class Todo extends Task {
