+++ src/main/java/duke/ui/Ui.java
@@ -1,5 +1,7 @@
+package duke.ui;
 import java.util.ArrayList;
 import java.util.Scanner;
+import duke.task.Task;
 
 public class Ui {
     private Scanner scanner;
